requirements.txt
cffi==1.5.2
cryptography==1.2.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-cas-client==1.2.0
django-guardian==1.4.6
django-toolbelt==0.0.1
google-api-python-client==1.5.0
gunicorn==19.2.0
httplib2==0.9.2
idna==2.0
ipaddress==1.0.16
oauth2client==2.0.1
Pillow==2.7.0
psycopg2==2.5.4
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.10
pycrypto==2.6.1
pyOpenSSL==0.14
python-dateutil==2.4.2
pytz==2015.4
requests==2.9.1
rsa==3.4
simplejson==3.6.5
six==1.10.0
static3==0.5.1
uritemplate==0.6
